# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Mercadeiro API
  description: >
    API do Mercadeiro — app de lista de compras domiciliar (Despensa → Carrinho →
    Finalizar Compra → Analítico). Documenta os endpoints da Fase 1 (API core).
    Regras de negócio referenciadas (RN-xx) e casos de uso (UC-xx) estão detalhados
    em `mercadeiro-regras-de-negocio.md`, na raiz do monorepo.


    **Autenticação:** todas as rotas exigem `Authorization: Bearer <token>`,
    exceto `/health`. Token único de aplicação (app single-user, sem contas —
    ver decisão em `mercadeiro-regras-de-negocio.md` §5.8), configurado em
    `API_BEARER_TOKEN` no `.env` do backend. Requisições sem token válido
    recebem `401 Unauthorized`.
  version: 0.1.0
servers:
  - url: http://localhost:3333
    description: Ambiente local de desenvolvimento
security:
  - bearerAuth: []
...
paths:
...
  /compras/atual:
    get:
      tags: [Compras]
      summary: Busca a Compra em andamento, se existir
      operationId: buscarCompraAtual
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "200":
          description: Compra em andamento
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Compra"
        "404":
          description: Não há Compra em andamento
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /compras:
    post:
      tags: [Compras]
      summary: Inicia uma sessão de compra, ou retoma a existente
      description: >
        RN-02.4/RN-05.5: só pode existir uma Compra "em andamento" por vez. Se já
        existir uma, retorna-a (status 200) em vez de criar outra. Ao criar uma
        nova (status 201), copia todos os ItemDespensa ativos para ItemCompra como
        `origem: PLANEJADO` e `comprado: false`, sem remover nada da Despensa
        (RN-02.3).
      operationId: criarOuRetomarCompra
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                supermercadoId:
                  type: string
                latitude:
                  type: number
                longitude:
                  type: number
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "200":
          description: Compra em andamento existente, retomada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Compra"
        "201":
          description: Nova Compra criada a partir da Despensa
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Compra"
        "400":
          $ref: "#/components/responses/BadRequest"
  /compras/{id}:
    parameters:
      - $ref: "#/components/parameters/CompraId"
    get:
      tags: [Compras]
      summary: Busca uma Compra por id, com seus itens
      operationId: buscarCompra
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "200":
          description: Compra encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Compra"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Compras]
      summary: Atualiza o supermercado e/ou a localização da Compra
      description: >
        RN-03.5: o supermercado pode ser trocado manualmente a qualquer momento
        antes de finalizar, sem afetar itens já marcados. RN-05.4: rejeitado se a
        Compra já estiver finalizada.
      operationId: atualizarCompra
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                supermercadoId:
                  type: string
                  nullable: true
                latitude:
                  type: number
                longitude:
                  type: number
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "200":
          description: Compra atualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Compra"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/CompraFinalizada"
  /compras/{id}/itens:
    parameters:
      - $ref: "#/components/parameters/CompraId"
    post:
      tags: [Compras]
      summary: Adiciona um item extra (não planejado) à Compra
      description: >
        UC-03: cria um ItemCompra com `origem: EXTRA`, `comprado: false` e sem
        preço. Se o produto já tiver um item extra nesta Compra, soma a
        quantidade em vez de duplicar. Rejeitado se a Compra já estiver
        finalizada (RN-05.4).
      operationId: adicionarItemExtra
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nome]
              properties:
                nome:
                  type: string
                  minLength: 1
                quantidade:
                  type: integer
                  minimum: 1
                  default: 1
                categoriaId:
                  type: string
                codigoBarras:
                  type: string
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "201":
          description: Item extra criado (ou quantidade somada a um existente)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemCompra"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/CompraFinalizada"
  /compras/{id}/itens/{itemId}:
    parameters:
      - $ref: "#/components/parameters/CompraId"
      - $ref: "#/components/parameters/ItemCompraId"
    patch:
      tags: [Compras]
      summary: Marca/desmarca um item como comprado, define preço, quantidade ou código lido
      description: >
        RN-03.1: um item só pode ser marcado como comprado (`comprado: true`) se
        já tiver `precoUnitario` definido ou se ele for informado nesta mesma
        requisição — caso contrário, 400. Desmarcar (`comprado: false`) mantém o
        preço salvo. Rejeitado se a Compra já estiver finalizada (RN-05.4).
      operationId: atualizarItemCompra
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Ao menos um campo deve ser informado
              properties:
                comprado:
                  type: boolean
                precoUnitario:
                  type: number
                  minimum: 0
                  exclusiveMinimum: true
                  nullable: true
                codigoBarrasLido:
                  type: string
                quantidade:
                  type: integer
                  minimum: 1
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "200":
          description: Item atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemCompra"
        "400":
          description: Dados inválidos, ou tentativa de marcar como comprado sem preço (RN-03.1)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/CompraFinalizada"
    delete:
      tags: [Compras]
      summary: Remove um item da Compra
      operationId: removerItemCompra
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "204":
          description: Item removido
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/CompraFinalizada"
  /compras/{id}/itens/{itemId}/quantidade:
    parameters:
      - $ref: "#/components/parameters/CompraId"
      - $ref: "#/components/parameters/ItemCompraId"
    patch:
      tags: [Compras]
      summary: Ajusta a quantidade de um item da Compra via incremento/decremento
      description: >
        Mesmo padrão do stepper da Despensa: um delta que levaria a quantidade a
        0 ou menos remove o item diretamente.
      operationId: ajustarQuantidadeItemCompra
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [delta]
              properties:
                delta:
                  type: integer
                  description: Valor a somar à quantidade atual; não pode ser zero
                  example: -1
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "200":
          description: Item removido ou item com quantidade atualizada
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ItemCompra"
                  - $ref: "#/components/schemas/RemocaoConfirmada"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/CompraFinalizada"
  /compras/{id}/finalizar:
    parameters:
      - $ref: "#/components/parameters/CompraId"
    post:
      tags: [Compras]
      summary: Finaliza a Compra, aplicando os efeitos na Despensa e no histórico de preço
      description: >
        RN-05.1 a RN-05.4, operação transacional (tudo ou nada):
        * Item planejado comprado → remove o ItemDespensa correspondente e grava PrecoHistorico.
        * Item extra comprado → grava PrecoHistorico, não cria ItemDespensa.
        * Item planejado não comprado → mantém/recria o ItemDespensa, sem preço.
        * Item extra não comprado → descartado silenciosamente (RN-05.1).

        Depois de finalizada, a Compra é somente leitura: chamadas repetidas a
        este endpoint, ou qualquer edição de itens, retornam 409.
      operationId: finalizarCompra
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "200":
          description: Compra finalizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Compra"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/CompraFinalizada"
...
components:
  parameters:
...
    CompraId:
      name: id
      in: path
      required: true
      schema:
        type: string
...
    ItemCompraId:
      name: itemId
      in: path
      required: true
      schema:
        type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        Token único de aplicação (API_BEARER_TOKEN no .env do backend).
        Exigido em todas as rotas, exceto /health.
  responses:
...
    Unauthorized:
      description: Token de autenticação ausente ou inválido
      headers:
        WWW-Authenticate:
          schema:
            type: string
            example: Bearer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: Dados inválidos
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    CompraFinalizada:
      description: Compra já finalizada é somente leitura (RN-05.4)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          description: >
            Detalhes adicionais, quando aplicável — formato varia por origem:
            array de issues do Zod em erros de validação, objeto de metadados
            do Prisma em conflitos de unicidade, ou ausente.
    RemocaoConfirmada:
      type: object
      required: [removido, id]
      properties:
        removido:
          type: boolean
          enum: [true]
        id:
          type: string
    Categoria:
      type: object
      required: [id, nome, ordem, createdAt, updatedAt]
      properties:
        id:
          type: string
        nome:
          type: string
        ordem:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Produto:
      type: object
      required: [id, nome, categoriaId, createdAt, updatedAt]
      properties:
        id:
          type: string
        nome:
          type: string
        codigoBarras:
          type: string
          nullable: true
        categoriaId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProdutoComCategoria:
      allOf:
        - $ref: "#/components/schemas/Produto"
        - type: object
          properties:
            categoria:
              $ref: "#/components/schemas/Categoria"
    Supermercado:
      type: object
      required: [id, nome, createdAt, updatedAt]
      properties:
        id:
          type: string
        nome:
          type: string
        endereco:
          type: string
          nullable: true
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    OrigemItemCompra:
      type: string
      enum: [PLANEJADO, EXTRA]
      description: >
        Definida na criação do item e não muda durante a compra.
        🏠 planejado (veio da Despensa) · ⚡ extra (adicionado durante a compra).
    ItemCompra:
      type: object
      required: [id, compraId, produtoId, quantidade, comprado, origem, produto, createdAt, updatedAt]
      properties:
        id:
          type: string
        compraId:
          type: string
        produtoId:
          type: string
        quantidade:
          type: integer
          minimum: 1
        precoUnitario:
          type: string
          nullable: true
          description: Valor decimal serializado como string (ex. "4.50"). Nulo até o item ser comprado.
          example: "4.50"
        comprado:
          type: boolean
        origem:
          $ref: "#/components/schemas/OrigemItemCompra"
        codigoBarrasLido:
          type: string
          nullable: true
        produto:
          $ref: "#/components/schemas/ProdutoComCategoria"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    StatusCompra:
      type: string
      enum: [EM_ANDAMENTO, FINALIZADA]
    Compra:
      type: object
      required: [id, status, itens, iniciadaEm, createdAt, updatedAt]
      properties:
        id:
          type: string
        status:
          $ref: "#/components/schemas/StatusCompra"
        supermercadoId:
          type: string
          nullable: true
        supermercado:
          type: object
          nullable: true
          allOf:
            - $ref: "#/components/schemas/Supermercado"
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        iniciadaEm:
          type: string
          format: date-time
        finalizadaEm:
          type: string
          format: date-time
          nullable: true
        itens:
          type: array
          items:
            $ref: "#/components/schemas/ItemCompra"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...